requirements.txt
tensorflow==2.4.1
opencv-python==4.5.1.48
Pillow==8.1.2
pyzbar==0.1.8
streamlit==0.78.0
tensorboard==2.4.1